Bugs fix register controller and test in postman

+++ src/utils/cloudinary.js
@@ -1,5 +1,7 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+dotenv.config();
 
 cloudinary.config({
   cloud_name: process.env.CLOUD_NAME,
@@ -9,13 +11,22 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
-    if (!uploadOnCloudinary) return null;
+    if (!localFilePath) return null;
+
+    // Upload file
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
-    console.log("File is uploaded on cloudinary", response.url);
-  } catch (error) {
+
+    // Remove from local after upload
     fs.unlinkSync(localFilePath);
+
+    return response;
+  } catch (error) {
+    if (localFilePath && fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    return null;
   }
 };
